+++ database/migrations/2020_10_01_195234_create_categories_table.php
@@ -16,7 +16,7 @@ class CreateCategoriesTable extends Migration
         Schema::create('categories', function (Blueprint $table) {
             $table->id();
             $table->string("name",100);
-            $table->integer("sub_category");
+            $table->unsignedBigInteger("sub_category");
             $table->timestamps();
         });
     }
